+++ src/commands/help.js
@@ -1,0 +1,4 @@
+console.log(
+  'Commands list:',
+  'up, '
+)
