--- src/commands/test.js
@@ -1,6 +1,0 @@
-import os from 'os';
-
-const params = 'EOL';
-
-console.log(JSON.stringify([os].params))
-console.log(process.argv.slice(2)[0])
