+++ src/core/cli.js
